fixed: refactored service classes to use invalidIdError from errorHandler
