+++ src/Services/locationService.ts
@@ -2,9 +2,7 @@ import { locationRepo } from '../Repositories/locationRepository';
 import { Location } from '../Entities/Location';
 import { LocationRequestDTO, LocationResponseDTO } from '../DTO/locationDTO';
 import { BaseError } from '../Utils/BaseError';
-import { Result, ApiResponse, failed, generateStatusCode } from '../Utils/errorHandler';
-
-const invalidIdError = new Error('No location with that id');
+import { Result, ApiResponse, failed, generateStatusCode, invalidIdError } from '../Utils/errorHandler';
 
 export const createLocation = async (locationRequestDTO: LocationRequestDTO): Promise<Result<ApiResponse, BaseError>> => {
     try {
@@ -31,7 +29,7 @@ export const getLocationById = async (id: number): Promise<Result<ApiResponse, B
     try {
         const response = await locationRepo.findOneByID(id);
         if (!response) {
-            return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+            return failed(invalidIdError('location'), await generateStatusCode(invalidIdError('location').message));
         }
         return success(response);
 
@@ -55,7 +53,7 @@ export const deleteLocation = async (locationId: number): Promise<Result<ApiResp
         const locationDB = await locationRepo.findOneByID(locationId);
 
         if (!locationDB) {
-            return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+            return failed(invalidIdError('location'), await generateStatusCode(invalidIdError('location').message));
         }
         const response = await locationRepo.remove(locationDB);
         return success(response);
