+++ src/Services/nameService.ts
@@ -2,9 +2,7 @@ import { NameRequestDTO, NameResponseDTO } from '../DTO/nameDTO';
 import { Name } from '../Entities/Name';
 import { nameRepo } from '../Repositories/nameRepository';
 import { BaseError } from '../Utils/BaseError';
-import { Result, ApiResponse, failed, generateStatusCode } from '../Utils/errorHandler';
-
-const invalidIdError = new Error('No name with that id');
+import { Result, ApiResponse, failed, generateStatusCode, invalidIdError } from '../Utils/errorHandler';
 
 export const createName = async (nameRequestDTO: NameRequestDTO): Promise<Result<ApiResponse, BaseError>> => {
   try {
@@ -22,7 +20,7 @@ export const getNameByID = async (id: number): Promise<Result<ApiResponse, BaseE
     const response = await nameRepo.findOneByID(id);
 
     if (!response) {
-      return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+      return failed(invalidIdError('name'), await generateStatusCode(invalidIdError('name').message));
     }
 
     return success(response);
@@ -58,7 +56,7 @@ export const deleteNameByID = async (id: number): Promise<Result<ApiResponse, Ba
     const nameDB = await nameRepo.findOneByID(id);
 
     if (!nameDB) {
-      return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+      return failed(invalidIdError('name'), await generateStatusCode(invalidIdError('name').message));
     }
     const response = await nameRepo.remove(nameDB);
     return success(response);
