+++ src/Services/parentService.ts
@@ -1,10 +1,8 @@
 import { parentRepo } from '../Repositories/parentRepository';
 import { Parent } from '../Entities/Parent';
 import { ParentRequestDTO, ParentResponseDTO } from '../DTO/parentDTO';
-import { Result, ApiResponse, failed, generateStatusCode } from '../Utils/errorHandler';
+import { Result, ApiResponse, failed, generateStatusCode, invalidIdError } from '../Utils/errorHandler';
 import { BaseError } from '../Utils/BaseError';
-
-const invalidIdError = new Error('No parent with that id');
 
 export const createParent = async (parentRequestDTO: ParentRequestDTO) => {
     try {
@@ -34,7 +32,7 @@ export const getParentById = async (id: number) => {
         const response = await parentRepo.findOneByID(id);
 
         if (!response) {
-            return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+            return failed(invalidIdError('parent'), await generateStatusCode(invalidIdError('parent').message));
         }
 
         return success(response);
@@ -59,7 +57,7 @@ export const deleteParent = async (parentId: number) => {
         const parentDB = await parentRepo.findOneByID(parentId);
 
         if (!parentDB) {
-            return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+            return failed(invalidIdError('parent'), await generateStatusCode(invalidIdError('parent').message));
         }
         const response = await parentRepo.remove(parentDB);
         return success(response);
