+++ src/Services/userService.ts
@@ -3,16 +3,14 @@ import { User } from '../Entities/User';
 import { UserRequestDTO, UserResponseDTO } from '../DTO/userDTO';
 import { getRoleById } from './roleService';
 import { Role } from '../Entities/Role';
-import { Result, ApiResponse, failed, generateStatusCode } from '../Utils/errorHandler';
+import { Result, ApiResponse, failed, generateStatusCode, invalidIdError } from '../Utils/errorHandler';
 import { BaseError } from '../Utils/BaseError';
-
-const invalidIdError = new Error('No user with that id');
 
 export const createUser = async (UserRequestDTO: UserRequestDTO) => {
     try {
         const role = await getRoleById(3) as unknown as Role;
 if(!role){
-    return failed(new Error('No role with that id'), await generateStatusCode(invalidIdError.message));
+    return failed(invalidIdError('role'), await generateStatusCode(invalidIdError('role').message));
 }
         UserRequestDTO.roles = [];
         UserRequestDTO.roles.push(role);
@@ -29,7 +27,7 @@ export const getUserByID = async (id: number) => {
     try {
         const response = await userRepo.findOneByID(id);
         if (!response) {
-            return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+            return failed(invalidIdError('user'), await generateStatusCode(invalidIdError('user').message));
         }
         return success(response);
     } catch (err) {
@@ -56,7 +54,7 @@ export const updateUser = async (userDTO: UserRequestDTO, email: string) => {
 
     const savedUser = await userRepo.save(userDB);
     if (!savedUser) {
-        return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+        return failed(invalidIdError('user'), await generateStatusCode(invalidIdError('user').message));
     }   
 
     return success(savedUser);
@@ -73,7 +71,7 @@ export const deleteUserByID = async (id: number) => {
         const response = await userRepo.findOneByID(id);
 
         if (!response || !response.userActive) {
-            return failed(invalidIdError, await generateStatusCode(invalidIdError.message));
+            return failed(invalidIdError('user'), await generateStatusCode(invalidIdError('user').message));
         }
     
         response.userActive = false;
